Renamed reader file to better reflect its content.
